+++ src/Form/ListeProjetParticipantsType.php
@@ -28,6 +28,7 @@ class ListeProjetParticipantsType extends AbstractType
             ])
             ->add('submit', SubmitType::class, [
                 'label' => 'Mettre à jour',
+                'attr' => ['class' => 'btn btn-success'],
             ])
             ->addEventListener(FormEvents::SUBMIT, [$this, 'putAllParticipantOnProjet'])
         ;